fix: add ability to delete deps.zip file when LIVE_DEPENDENCIES config is empty (#5)
The job runner would only run successfully if there was an entry in the LIVE_DEPENDENCIES config variable. This meant that atleast one entry should exist in that variable and that it was impossible to delete all of the live dependencies. This change makes it so the runner would delete the deps.zip file if LIVE_DEPENDENCIES config variable is empty.