feat: merge all livedeps scripts into one file, update docs, add init job and lots of refactoring (#7)
Here are the documentation changes made in this PR:
  1. Added instructions  in docs to run the livedeps init job after installing plugin.
  2. Renamed LIVE_DEPENDENCIES config to LIVEDEPS for consistency.
  3. Removed dev from allowed commands since the plugin will not work for tutor dev.

All relevant scripts are merged into a single livedeps script which is placed at /openedx/bin/. This is already in PATH which means we can call the livedeps script directly.

This change also includes lots of refactoring and additional comments in code.